updated the functionalities and made changes in Landing, Signup, CartPage and edit profile pages
Landing page- 1. Updated the header as per the new suggestions in both mobile and desktop view
2. fixed the footer to the end of the page i all the conditions

Signup page: 1.fixed the issue in storing the user name
(previously it was storing the email in the username field now fixed that to correctly store the user name

Cart Page: 1.Restricted the quantity increase button based on the newly added stock availability feature

Edit Profile Page: 1. Made the ui responsive and as per the requirement
2. Added all the required functionalities like edit profile, save changes, discard changes, change password
3. fixed all the issues in updating the changes correctly and dynamically displaying in the respective parts of the webpage without refreshing the page.
4.Made the change password model as a pop up over the edit password page and made sure that the security matching for the existing password works fine